[Assignment6]Add offset, rotation and scaling

+++ Assignment6/Transformations.cs
@@ -8,26 +8,26 @@ namespace Assignment6
 {
     class Transformations
     {
-        private double[,] matrix = new double[4, 4];
+        private float[,] matrix = new float[4, 4];
 
-        public double[,] Matrix { get { return matrix; } }
+        public float[,] Matrix { get { return matrix; } }
 
         public Transformations()
         {
             matrix = Identity().matrix;
         }
 
-        public Transformations(double[,] matrix)
+        public Transformations(float[,] matrix)
         {
             this.matrix = matrix;
         }
 
         public static Transformations RotateX(double angle)
         {
-            double cos = Math.Cos(angle);
-            double sin = Math.Sin(angle);
+            float cos = (float)Math.Cos(angle);
+            float sin = (float)Math.Sin(angle);
             return new Transformations(
-                new double[,]
+                new float[,]
                 {
                     { 1, 0, 0, 0 },
                     { 0, cos, -sin, 0 },
@@ -38,10 +38,10 @@ namespace Assignment6
 
         public static Transformations RotateY(double angle)
         {
-            double cos = Math.Cos(angle);
-            double sin = Math.Sin(angle);
+            float cos = (float)Math.Cos(angle);
+            float sin = (float)Math.Sin(angle);
             return new Transformations(
-                new double[,]
+                new float[,]
                 {
                     { cos, 0, sin, 0 },
                     { 0, 1, 0, 0 },
@@ -52,10 +52,10 @@ namespace Assignment6
 
         public static Transformations RotateZ(double angle)
         {
-            double cos = Math.Cos(angle);
-            double sin = Math.Sin(angle);
+            float cos = (float)Math.Cos(angle);
+            float sin = (float)Math.Sin(angle);
             return new Transformations(
-                new double[,]
+                new float[,]
                 {
                     { cos, -sin, 0, 0 },
                     { sin, cos, 0, 0 },
@@ -66,15 +66,15 @@ namespace Assignment6
 
         public static Transformations RotateLine(Line line, double angle)
         {
-            double cos = Math.Cos(angle);
-            double sin = Math.Sin(angle);
-            double l = Math.Sign(line.pos2.X - line.pos1.X);
-            double m = Math.Sign(line.pos2.Y - line.pos1.Y);
-            double n = Math.Sign(line.pos2.Z - line.pos1.Z);
-            double length = Math.Sqrt(l * l + m * m + n * n);
+            float cos = (float)Math.Cos(angle);
+            float sin = (float)Math.Sin(angle);
+            float l = Math.Sign(line.pos2.X - line.pos1.X);
+            float m = Math.Sign(line.pos2.Y - line.pos1.Y);
+            float n = Math.Sign(line.pos2.Z - line.pos1.Z);
+            float length = (float)Math.Sqrt(l * l + m * m + n * n);
             l /= length; m /= length; n /= length;
             return new Transformations(
-                new double[,]
+                new float[,]
                 {
                    { l * l + cos * (1 - l * l),   l * (1 - cos) * m + n * sin,   l * (1 - cos) * n - m * sin,  0  },
                    { l * (1 - cos) * m - n * sin,   m * m + cos * (1 - m * m),    m * (1 - cos) * n + l * sin,  0 },
@@ -87,10 +87,10 @@ namespace Assignment6
         public static Transformations Scale(double fx, double fy, double fz)
         {
             return new Transformations(
-                new double[,] {
-                    { fx, 0, 0, 0 },
-                    { 0, fy, 0, 0 },
-                    { 0, 0, fz, 0 },
+                new float[,] {
+                    { (float)fx, 0, 0, 0 },
+                    { 0, (float)fy, 0, 0 },
+                    { 0, 0, (float)fz, 0 },
                     { 0, 0, 0, 1 }
                 });
         }
@@ -98,19 +98,19 @@ namespace Assignment6
         public static Transformations Translate(double dx, double dy, double dz)
         {
             return new Transformations(
-                new double[,]
+                new float[,]
                 {
                     { 1, 0, 0, 0 },
                     { 0, 1, 0, 0 },
                     { 0, 0, 1, 0 },
-                    { dx, dy, dz, 1 },
+                    { (float)dx, (float)dy, (float)dz, 1 },
                 });
         }
 
         public static Transformations Identity()
         {
             return new Transformations(
-                new double[,] {
+                new float[,] {
                     { 1, 0, 0, 0 },
                     { 0, 1, 0, 0 },
                     { 0, 0, 1, 0 },
@@ -157,7 +157,7 @@ namespace Assignment6
         public static Transformations PerspectiveProjection()
         {
             return new Transformations(
-                new double[,] {
+                new float[,] {
                     { 1, 0, 0, 0 },
                     { 0, 1, 0, 0 },
                     { 0, 0, 0, 2 },
@@ -167,7 +167,7 @@ namespace Assignment6
 
         public static Transformations operator *(Transformations t1, Transformations t2)
         {
-            double[,] matrix = new double[4, 4];
+            float[,] matrix = new float[4, 4];
             for (int i = 0; i < 4; ++i)
                 for (int j = 0; j < 4; ++j)
                 {
